revrote instruction parsing

+++ test_c/main.c
@@ -1,6 +1,6 @@
-short main()
+int main()
 {
-    short a = 10;
-    short b = 20;
-    short c = a+b;
+    int a = 10;
+    int b = 20;
+    int c = a+b;
 }
